+++ scripts/analyse_missions.py
@@ -6,30 +6,19 @@ import logging
 
 from rich.progress import track
 
-from scripts.constants import BASE_PATH, CONFIG
-from src.mission.mission import analyse_mission
-from src.mission.utils import mission_dirs_in_dir
+from scripts.constants import PATHS
+from src.mission.mission import analyse_single_mission
+from src.mission.utils import map_name_from_mission_dir_path, mission_dirs_in_dir
 from src.utils import configure_logging, pretty_iterable_of_str
 from static_data import au_mission_overrides, in_game_data
 from static_data.map_index import MAP_INDEX
 
 LOGGER = logging.getLogger(__name__)
-PATHS = {
-    "AU_MAPS_DIR": (
-        BASE_PATH / CONFIG["AU_SOURCE_DIR_RELATIVE"] / "A3A/addons/maps"
-    ).resolve(),
-    "DATA_DIR": (BASE_PATH / CONFIG["INTERMEDIATE_DATA_DIR_RELATIVE"]).resolve(),
-}
 
 
 def analyse_missions() -> None:
     """Analyse all missions."""
     configure_logging()
-    for name, path in PATHS.items():
-        if not path.is_dir():
-            err_msg = f"{name}: {path} not found"
-            raise RuntimeError(err_msg)
-
     mission_dirs = sorted(
         d
         for d in mission_dirs_in_dir(PATHS["AU_MAPS_DIR"])
@@ -48,7 +37,13 @@ def analyse_missions() -> None:
     PATHS["DATA_DIR"].mkdir(parents=True, exist_ok=True)
     analysed_map_names = set()
     for mission_dir in track(mission_dirs, description="Analysing missions..."):
-        map_name = analyse_mission(mission_dir)
+        map_name = map_name_from_mission_dir_path(mission_dir)
+        analyse_single_mission(
+            au_map_dir=mission_dir,
+            map_index=MAP_INDEX,
+            grad_meh_map_dir=PATHS["GRAD_MEH_DIR"] / map_name,
+            export_dir=PATHS["DATA_DIR"],
+        )
         analysed_map_names.add(map_name)
 
     log_msg = (