+++ src/mission/marker.py
@@ -6,24 +6,13 @@ from typing import TYPE_CHECKING, Self
 
 from attrs import define
 
-from src.position_2d import Position2D
+from src.mission.position_2d import Position2D
 
 if TYPE_CHECKING:
     from src.types_ import DictNode
 
 
-RELEVANT_MARKER_PREFIXES = {
-    # case-insensitive
-    "airport",
-    "factory",
-    "milbase",
-    "outpost",
-    "resource",
-    "seaport",
-}
-
-
-@define
+@define(kw_only=True, frozen=True)
 class Marker:
     """Represents a map marker."""